pyproject.toml
[project]
name = "question-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.121.3",
    "pydantic-settings>=2.11.0",
    "pymongo>=4.15.4",
    "uvicorn[standard]>=0.38.0",
]
